app: return double not int from GimpPickable::get_opacity_at()
and fix GimpLayer's impl to honor the mask also for layers without
alpha.

+++ app/core/gimppickable.h
@@ -46,7 +46,7 @@ struct _GimpPickableInterface
                                              gint          y,
                                              const Babl   *format,
                                              gpointer      pixel);
-  gint            (* get_opacity_at)        (GimpPickable *pickable,
+  gdouble         (* get_opacity_at)        (GimpPickable *pickable,
                                              gint          x,
                                              gint          y);
 };
@@ -69,7 +69,7 @@ gboolean        gimp_pickable_get_color_at          (GimpPickable *pickable,
                                                      gint          x,
                                                      gint          y,
                                                      GimpRGB      *color);
-gint            gimp_pickable_get_opacity_at        (GimpPickable *pickable,
+gdouble         gimp_pickable_get_opacity_at        (GimpPickable *pickable,
                                                      gint          x,
                                                      gint          y);
